Remove unnecessary domains component
